git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/branches/issue-clo-462@15491 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/data/DynamicRecordBuffer.java
@@ -445,7 +445,12 @@ public class DynamicRecordBuffer {
     		if (diskSlot != null) {
     			return diskSlot;
     		} else {
-    			obsoleteTempFiles.removeFirst();
+    			obsoleteTempFiles.removeFirst(); //removes obsoleteTempFile
+    			try {
+    				obsoleteTempFile.close();
+    			} catch (IOException e) {
+    				log.warn("Failed to close temp file.", e);
+    			}
     		}
     	}
     	
@@ -560,11 +565,15 @@ public class DynamicRecordBuffer {
 			try {
 				fullFileBuffers = null;
 		        emptyFileBuffers = null;
-				tempFileChannel.close();
-			} finally {
-		        if (!tempFile.delete()) {
-		        	log.warn("Failed to delete temp file: " + tempFile.getAbsolutePath());
+		        if (tempFileChannel != null) {
+		        	tempFileChannel.close();
 		        }
+			} finally {
+				if (tempFile != null) {
+			        if (!tempFile.delete()) {
+			        	log.warn("Failed to delete temp file: " + tempFile.getAbsolutePath());
+			        }
+				}
 			}
 		}
